requirements.txt
Django==3.2
python-decouple==3.6
djangorestframework==3.13.1
django-filter==22.1
requests=2.28.1
djangorestframework-simplejwt==5.2.0
